Document Debian's decision to disable modelines by default
Modelines have historically been a source of vulnerabilities in Vim.
As long as it remains a "blacklist suspected/proven dangerous options"
instead of a "whitelist allowed options" piece of functionality,
Debian's system-wide vimrc will maintain this setting.  As such, the
documentation needs to be updated to reflect the induced behavior.

Closes: #472522
Signed-off-by: James McCoy <jamessan@debian.org>